refactor: decompose window manager into lifecycle modules
Split window-manager.ts into focused modules for guest webview security,
navigation guards, crash handlers, window positioning, spell-check menu, and
dev auto-updater stubs.

Keep the shared createBrowserWindow factory, multi-window registry wiring, and
plugin panel hardening behavior unchanged. window-manager.ts now delegates
security and crash detection to the extracted attach* helpers.

Preserve existing comments on security rationale, Sentry noise filtering, and
off-screen window placement.
