+++ src/main/app-lifecycle/dev-auto-updater-stubs.ts
@@ -1,0 +1,39 @@
+import { ipcMain } from 'electron';
+import { logger } from '../utils/logger';
+
+// Track if stub handlers have been registered (module-level to persist across createWindow calls)
+let devStubsRegistered = false;
+
+/**
+ * Registers stub IPC handlers for auto-updater in development mode.
+ * These provide helpful error messages instead of silent failures.
+ * Uses a module-level flag to ensure handlers are only registered once.
+ */
+export function registerDevAutoUpdaterStubs(): void {
+	// Only register once - prevents duplicate handler errors if createWindow is called multiple times
+	if (devStubsRegistered) {
+		logger.debug('Auto-updater stub handlers already registered, skipping', 'Window');
+		return;
+	}
+
+	ipcMain.handle('updates:download', async () => {
+		return {
+			success: false,
+			error: 'Auto-update is disabled in development mode. Please check update first.',
+		};
+	});
+
+	ipcMain.handle('updates:install', async () => {
+		logger.warn('Auto-update install called in development mode', 'AutoUpdater');
+	});
+
+	ipcMain.handle('updates:getStatus', async () => {
+		return { status: 'idle' as const };
+	});
+
+	ipcMain.handle('updates:checkAutoUpdater', async () => {
+		return { success: false, error: 'Auto-update is disabled in development mode' };
+	});
+
+	devStubsRegistered = true;
+}
