+++ src/main/app-lifecycle/window-manager.ts
@@ -10,8 +10,7 @@
  * `createSecondaryWindow` callers around that factory.
  */
 
-import path from 'path';
-import { BrowserWindow, Menu, ipcMain, screen } from 'electron';
+import { BrowserWindow } from 'electron';
 import type Store from 'electron-store';
 import type { SettingsStoreInterface, WindowState } from '../stores/types';
 import type { WindowState as SharedWindowState } from '../../shared/window-types';
@@ -23,217 +22,12 @@ import type { WindowRegistry } from '../window-registry';
 import { saveWindowState, WINDOW_STATE_SAVE_DEBOUNCE_MS } from '../window-state-persistence';
 import { debounce } from '../utils/debounce';
 import { isWebContentsAvailable } from '../utils/safe-send';
-import { isAllowedBrowserTabPartition } from '../../shared/browserTabPartition';
-import { blocksSubframeNavigation } from '../../shared/plugins/panel-navigation';
-import {
-	isPluginPanelPartition,
-	isAllowedPluginPanelAttachment,
-} from '../../shared/plugins/panel-host';
-import {
-	hardenPluginPanelWebPreferences,
-	hardenPluginPanelSession,
-	attachPluginPanelGuestSecurity,
-	isPluginPanelSession,
-	type PluginPanelGuestContents,
-} from '../plugins/plugin-panel-host';
-
-// `file:` is allowed so users can open local HTML they just generated
-// (Plotly dashboards, etc.) inside Maestro instead of bouncing to the system
-// browser. The webview is still hardened (sandbox, no node, webSecurity true)
-// and only renders content the user explicitly opens.
-const ALLOWED_BROWSER_TAB_EMBED_PROTOCOLS = new Set(['http:', 'https:', 'file:']);
-const ALLOWED_BROWSER_TAB_ABOUT_URLS = new Set(['about:blank']);
-const ALLOWED_APP_PERMISSIONS = new Set(['clipboard-read', 'clipboard-sanitized-write']);
-
-/** Sentry severity levels */
-type SentrySeverityLevel = 'fatal' | 'error' | 'warning' | 'log' | 'info' | 'debug';
-
-/** Sentry module type for crash reporting */
-interface SentryModule {
-	captureMessage: (
-		message: string,
-		captureContext?: { level?: SentrySeverityLevel; extra?: Record<string, unknown> }
-	) => string;
-}
-
-/** Cached Sentry module reference */
-let sentryModule: SentryModule | null = null;
-
-type BrowserTabWebPreferences = Record<string, unknown> & {
-	partition?: string;
-	preload?: string;
-	nodeIntegration?: boolean;
-	nodeIntegrationInSubFrames?: boolean;
-	contextIsolation?: boolean;
-	sandbox?: boolean;
-	webSecurity?: boolean;
-	allowRunningInsecureContent?: boolean;
-};
-
-interface BrowserTabGuestContents {
-	getType?: () => string;
-	setWindowOpenHandler: (
-		handler: ({ url }: { url: string }) => { action: 'deny' | 'allow' }
-	) => void;
-	on(
-		event: 'will-navigate' | 'will-redirect',
-		handler: (event: { preventDefault: () => void }, url: string) => void
-	): void;
-	on(event: string, handler: (...args: any[]) => void): void;
-	executeJavaScript(code: string): Promise<unknown>;
-	// Privileged Electron paste: bypasses the web-facing `clipboard-read`
-	// permission that the permission handler denies to webviews (issue #1063).
-	paste(): void;
-}
-
-function isAllowedBrowserTabUrl(rawUrl: string): boolean {
-	if (ALLOWED_BROWSER_TAB_ABOUT_URLS.has(rawUrl)) return true;
-
-	try {
-		return ALLOWED_BROWSER_TAB_EMBED_PROTOCOLS.has(new URL(rawUrl).protocol);
-	} catch {
-		return false;
-	}
-}
-
-function hardenBrowserTabWebPreferences(webPreferences: BrowserTabWebPreferences): void {
-	delete webPreferences.preload;
-	delete (webPreferences as Record<string, unknown>).preloadURL;
-
-	webPreferences.nodeIntegration = false;
-	webPreferences.nodeIntegrationInSubFrames = false;
-	webPreferences.contextIsolation = true;
-	webPreferences.sandbox = true;
-	webPreferences.webSecurity = true;
-	webPreferences.allowRunningInsecureContent = false;
-}
-
-function attachBrowserTabGuestSecurity(guestContents: BrowserTabGuestContents): void {
-	const denyBrowserTabNavigation = (
-		eventName: 'will-navigate' | 'will-redirect',
-		event: { preventDefault: () => void },
-		url: string
-	) => {
-		if (isAllowedBrowserTabUrl(url)) return;
-
-		event.preventDefault();
-		logger.warn(`Blocked browser-tab ${eventName}: ${url}`, 'Window', {
-			url,
-			type: guestContents.getType?.() ?? 'unknown',
-		});
-	};
-
-	guestContents.setWindowOpenHandler(({ url }) => {
-		logger.warn(`Blocked browser-tab popup: ${url}`, 'Window', {
-			url,
-			type: guestContents.getType?.() ?? 'unknown',
-		});
-		return { action: 'deny' };
-	});
-
-	guestContents.on('will-navigate', (event, url) => {
-		denyBrowserTabNavigation('will-navigate', event, url);
-	});
-
-	guestContents.on('will-redirect', (event, url) => {
-		denyBrowserTabNavigation('will-redirect', event, url);
-	});
-}
-
-/**
- * Reports a crash event to Sentry from the main process.
- * Lazily loads Sentry to avoid module initialization issues.
- */
-async function reportCrashToSentry(
-	message: string,
-	level: SentrySeverityLevel,
-	extra?: Record<string, unknown>
-): Promise<void> {
-	try {
-		if (!sentryModule) {
-			const sentry = await import('@sentry/electron/main');
-			sentryModule = sentry;
-		}
-		sentryModule.captureMessage(message, { level, extra });
-	} catch {
-		// Sentry not available (development mode or initialization failed)
-		logger.debug('Sentry not available for crash reporting', 'Window');
-	}
-}
-
-/** A display work-area rectangle in screen (DIP) coordinates. */
-type DisplayWorkArea = { x: number; y: number; width: number; height: number };
-
-/**
- * Centers a window of the given size inside a display's work area. The offset is
- * clamped to zero so a window larger than the work area still pins to its
- * top-left corner (its title bar) rather than spilling above/left of it.
- */
-function centerWithinWorkArea(
-	workArea: DisplayWorkArea,
-	width: number,
-	height: number
-): { x: number; y: number } {
-	return {
-		x: Math.round(workArea.x + Math.max(0, (workArea.width - width) / 2)),
-		y: Math.round(workArea.y + Math.max(0, (workArea.height - height) / 2)),
-	};
-}
-
-/**
- * Resolves the on-screen position for a window restored from saved bounds,
- * accounting for display-configuration changes between sessions.
- *
- * The saved bounds are validated against the *current* displays:
- * `screen.getDisplayMatching` returns the display the saved rectangle most
- * closely intersects (when the monitor that held the window has been unplugged
- * this falls back to the nearest remaining display), and we check whether the
- * window's title bar would actually be reachable on it. Two cases leave the
- * saved coordinates unusable and the window invisible:
- *   - the window was saved minimized (Windows reports bounds of -32000,-32000), or
- *   - the monitor it lived on has been removed.
- * In both cases the window is repositioned onto the primary display so it can
- * never spawn off-screen. When there is no saved position at all we return
- * undefined x/y so Electron places the window itself.
- */
-export function resolveVisibleWindowPosition(state: {
-	x?: number;
-	y?: number;
-	width: number;
-	height: number;
-}): { x?: number; y?: number } {
-	if (typeof state.x !== 'number' || typeof state.y !== 'number') {
-		return {};
-	}
-
-	const bounds = { x: state.x, y: state.y, width: state.width, height: state.height };
-
-	// Validate against the display the saved bounds most closely intersect. If
-	// that monitor is gone, getDisplayMatching returns the nearest remaining one
-	// and the reachability check below fails, triggering a reposition.
-	const matched = screen.getDisplayMatching(bounds);
-
-	// The window is reachable if the center of its title bar lands inside the
-	// matched display's work area, with a bottom margin so the title bar can't
-	// sit below the screen edge where it can't be grabbed.
-	const BOTTOM_MARGIN = 80;
-	const TITLE_BAR_SAMPLE_Y = 16; // approximate title-bar height (px)
-	const titleBar = { x: bounds.x + bounds.width / 2, y: bounds.y + TITLE_BAR_SAMPLE_Y };
-	const { x, y, width, height } = matched.workArea;
-	const isOnScreen =
-		titleBar.x >= x &&
-		titleBar.x <= x + width &&
-		titleBar.y >= y &&
-		titleBar.y <= y + height - BOTTOM_MARGIN;
-
-	if (isOnScreen) {
-		return { x: bounds.x, y: bounds.y };
-	}
-
-	// Off-screen (minimized sentinel or removed monitor): bring the window back
-	// onto the primary display so it can never spawn invisible.
-	return centerWithinWorkArea(screen.getPrimaryDisplay().workArea, bounds.width, bounds.height);
-}
+import { attachGuestWebviewSecurity } from './guest-webview-security';
+import { attachMainWindowNavigationGuards } from './main-window-navigation';
+import { attachSpellCheckContextMenu } from './spell-check-menu';
+import { attachWindowCrashHandlers } from './window-crash-handlers';
+import { registerDevAutoUpdaterStubs } from './dev-auto-updater-stubs';
+import { resolveVisibleWindowPosition } from './window-position';
 
 /** Bounds/state used to size and position a window at creation time. */
 interface WindowCreationBounds {
@@ -501,283 +295,14 @@ export function createWindowManager(deps: WindowManagerDependencies): WindowMana
 		// Navigation & Window Security Hardening
 		// ================================================================
 
-		// The plugin-panel preload lives next to the main preload bundle
-		// (dist/main/plugin-panel-preload.js, built by scripts/build-preload.mjs).
-		const pluginPanelPreloadPath = path.join(path.dirname(preloadPath), 'plugin-panel-preload.js');
-
-		// Restrict renderer-created webviews to the two sanctioned surfaces:
-		// browser tabs (persist:maestro-browser-session-*) and plugin panels
-		// (plugin:<id>, hardened by the plugin panel host). Anything else is
-		// blocked before the guest exists.
-		browserWindow.webContents.on('will-attach-webview', (event, webPreferences, params) => {
-			const src = typeof params.src === 'string' ? params.src : '';
-			const partition =
-				typeof webPreferences.partition === 'string' ? webPreferences.partition : '';
-
-			if (isPluginPanelPartition(partition)) {
-				// Per-plugin isolated surface: partition and document must belong
-				// to the SAME plugin, web prefs are forced (no Node, isolation,
-				// sandbox, broker-only preload), and the per-plugin session gets
-				// its protocol handler + egress/permission denial installed.
-				if (!isAllowedPluginPanelAttachment(partition, src)) {
-					event.preventDefault();
-					logger.warn(`Blocked unsafe plugin panel attachment: ${src || '<empty src>'}`, 'Window', {
-						src,
-						partition,
-					});
-					return;
-				}
-				hardenPluginPanelWebPreferences(
-					webPreferences as Record<string, unknown>,
-					pluginPanelPreloadPath
-				);
-				hardenPluginPanelSession(partition);
-				return;
-			}
-
-			hardenBrowserTabWebPreferences(webPreferences as BrowserTabWebPreferences);
-
-			if (!isAllowedBrowserTabUrl(src) || !isAllowedBrowserTabPartition(partition)) {
-				event.preventDefault();
-				logger.warn(`Blocked unsafe webview attachment: ${src || '<empty src>'}`, 'Window', {
-					src,
-					partition,
-				});
-			}
+		attachGuestWebviewSecurity(browserWindow, preloadPath);
+		attachMainWindowNavigationGuards(browserWindow, {
+			isDevelopment,
+			devServerUrl,
+			rendererProductionUrl,
+			entryUrl,
 		});
-
-		browserWindow.webContents.on('did-attach-webview', (_event, guestContents) => {
-			// Plugin panel guests get the panel lockdown ONLY: no popups, no
-			// navigation — and none of the browser-tab conveniences (shortcut
-			// forwarding, JS injection, privileged paste) may ever run inside
-			// plugin-controlled content.
-			const panelGuest = guestContents as unknown as PluginPanelGuestContents;
-			if (isPluginPanelSession(panelGuest.session)) {
-				attachPluginPanelGuestSecurity(panelGuest);
-				return;
-			}
-
-			attachBrowserTabGuestSecurity(guestContents as BrowserTabGuestContents);
-
-			// Forward app shortcuts from the webview guest process to the renderer.
-			// When a <webview> has focus, keyboard events are trapped in its guest
-			// Chromium process and never reach the renderer's window keydown handler.
-			//
-			// Strategy: inject a bubble-phase keydown listener into the guest page.
-			// After all page handlers have run, if the page did NOT call preventDefault
-			// on a Meta/Ctrl keystroke, it means the page doesn't use that shortcut —
-			// so we forward it to the app. If the page DID preventDefault, the page's
-			// shortcut takes precedence and we leave it alone.
-			const guest = guestContents as BrowserTabGuestContents;
-
-			// Intercept app shortcuts BEFORE Chromium's built-in handlers consume them.
-			// Some keys (e.g. Cmd+L for address bar focus) are handled by Chromium
-			// internally and never reach the injected JS listener below.
-			guest.on('before-input-event', (event, input) => {
-				if (!input.meta && !input.control && !input.alt) return;
-				if (input.type !== 'keyDown') return;
-				const k = input.key.toLowerCase();
-				// Cmd/Ctrl+V: drive paste through the trusted guest webContents API.
-				// Chromium's native paste needs the `clipboard-read` permission, which
-				// the permission handler denies to webviews as a security boundary, so
-				// native paste silently fails inside browser-tab form fields (issue
-				// #1063). guest.paste() is a privileged Electron call that bypasses
-				// that web-facing permission, mirroring the right-click Paste menu
-				// item (issue #1065).
-				const isPaste = (input.meta || input.control) && !input.alt && !input.shift && k === 'v';
-				if (isPaste) {
-					event.preventDefault();
-					guest.paste();
-					return;
-				}
-				// Let the remaining standard text-editing shortcuts pass through to
-				// the page. `f` is intentionally NOT in this list: Cmd+F must reach
-				// the renderer so the in-page find bar can open.
-				const isTextEditing =
-					(input.meta || input.control) && !input.alt && !input.shift && 'acxz'.includes(k);
-				const isRedo = (input.meta || input.control) && !input.alt && input.shift && k === 'z';
-				if (isTextEditing || isRedo) return;
-				event.preventDefault();
-				browserWindow.webContents.send('browser-tab:shortcutKey', {
-					key: input.key,
-					code: input.code,
-					meta: input.meta,
-					control: input.control,
-					alt: input.alt,
-					shift: input.shift,
-				});
-			});
-
-			// Capture-phase listener: intercepts app shortcuts BEFORE the page
-			// can handle them.  We preventDefault+stopPropagation so the page
-			// never sees the event, then forward it to the app via console.log.
-			const shortcutInjection = `(function(){
-				if(window.__maestroShortcutListenerInstalled)return;
-				window.__maestroShortcutListenerInstalled=true;
-				document.addEventListener('keydown',function(e){
-					var hasMod=e.metaKey||e.ctrlKey;
-					var hasAlt=e.altKey;
-					if(!hasMod&&!hasAlt)return;
-					var k=e.key.toLowerCase();
-					var te=hasMod&&!hasAlt&&!e.shiftKey&&'acxz'.indexOf(k)!==-1;
-					var re=hasMod&&!hasAlt&&e.shiftKey&&k==='z';
-					if(te||re)return;
-					e.preventDefault();
-					e.stopPropagation();
-					console.log('__MAESTRO_KEY__'+JSON.stringify({
-						key:e.key,code:e.code,
-						meta:e.metaKey,control:e.ctrlKey,
-						alt:e.altKey,shift:e.shiftKey
-					}));
-				},true);
-			})();`;
-			const injectShortcutListener = () => {
-				guest.executeJavaScript(shortcutInjection).catch(() => {});
-			};
-			guest.on('dom-ready', injectShortcutListener);
-			guest.on('did-navigate', injectShortcutListener);
-			// console-message args: (event, level, message, line, sourceId)
-			guest.on('console-message', (...args: unknown[]) => {
-				const message = typeof args[2] === 'string' ? args[2] : String(args[2] ?? '');
-				const prefix = '__MAESTRO_KEY__';
-				if (!message.startsWith(prefix)) return;
-				try {
-					const input = JSON.parse(message.slice(prefix.length));
-					browserWindow.webContents.send('browser-tab:shortcutKey', input);
-				} catch {
-					// Malformed message, ignore
-				}
-			});
-		});
-
-		// Subframe egress guard (backstop). App-window `srcDoc` subframes (the
-		// file-preview renderer) have no business navigating anywhere: a meta CSP
-		// cannot stop such a frame from navigating ITSELF to a remote URL and
-		// leaking data through it, so block any subframe navigation away from its
-		// initial document here (the top frame is handled by `will-navigate`
-		// below). Plugin panels are NOT subframes anymore — they are <webview>
-		// guests with their own webContents, locked down separately in
-		// attachPluginPanelGuestSecurity (did-attach-webview) — but this guard
-		// stays as defense in depth for any srcDoc frame in the app window.
-		browserWindow.webContents.on('will-frame-navigate', (event) => {
-			if (!blocksSubframeNavigation(event.isMainFrame, event.url)) return;
-			event.preventDefault();
-			logger.warn(`Blocked subframe navigation to: ${event.url}`, 'Window');
-		});
-
-		// Deny all popup/new-window requests — external links use IPC shell:openExternal
-		browserWindow.webContents.setWindowOpenHandler(({ url }) => {
-			logger.warn(`Blocked window.open request: ${url}`, 'Window');
-			return { action: 'deny' };
-		});
-
-		// Restrict navigation to the app itself — prevent renderer from navigating away.
-		// Both the dev-server URL and the renderer entry's file:// URL are constants
-		// for the lifetime of this window, so compute them once at setup time rather
-		// than on every navigation event. The production guard only allows the
-		// renderer entry HTML itself: a previous "directory prefix" check let any
-		// file inside the renderer dir through, which meant a stray <a href="foo.md">
-		// in chat output could resolve relative to index.html and unload the app to
-		// a non-existent bundle file.
-		// The dev server serves the app at its root. A previous guard allowed the
-		// ENTIRE dev origin through, which let any same-origin path (a game served
-		// by the dev server, or a stray relative <a href="game/"> in chat/markdown
-		// output) unload the app and take over the whole window. Page content
-		// belongs in a <webview> browser tab, never the top-level frame, so the dev
-		// guard is now as strict as production: only the app's own entry document
-		// (origin AND pathname) may load top-level. HMR/full-reloads target the same
-		// root URL and the renderer has no top-level URL routing, so this is safe.
-		// `allowedProdEntryUrl` is THIS window's exact entry URL (which carries the
-		// `?windowId=` query for secondary windows) so a programmatic reload to the
-		// same URL is allowed while any other path is still rejected.
-		const devEntryUrl = isDevelopment ? new URL(devServerUrl) : null;
-		const allowedDevOrigin = devEntryUrl ? devEntryUrl.origin : null;
-		const allowedDevPathname = devEntryUrl ? devEntryUrl.pathname || '/' : null;
-		const allowedProdOrigin = isDevelopment ? null : new URL(rendererProductionUrl).origin;
-		const allowedProdEntryUrl = isDevelopment ? null : entryUrl;
-		browserWindow.webContents.on('will-navigate', (event, url) => {
-			const parsedUrl = new URL(url);
-			if (isDevelopment) {
-				const pathname = parsedUrl.pathname || '/';
-				if (parsedUrl.origin === allowedDevOrigin && pathname === allowedDevPathname) return;
-			} else {
-				if (parsedUrl.origin === allowedProdOrigin && url === allowedProdEntryUrl) return;
-			}
-			event.preventDefault();
-			logger.warn(`Blocked navigation to: ${url}`, 'Window');
-		});
-
-		// Deny most browser permission requests (camera, mic, geolocation, etc.)
-		// Allow clipboard access for the app window only, never embedded browser tabs.
-		browserWindow.webContents.session.setPermissionRequestHandler(
-			(webContents, permission, callback) => {
-				const contentsType = webContents?.getType?.();
-				const isAppWindow = contentsType === 'window';
-
-				if (isAppWindow && ALLOWED_APP_PERMISSIONS.has(permission)) {
-					callback(true);
-				} else {
-					if (contentsType === 'webview') {
-						logger.warn(`Blocked browser-tab permission request: ${permission}`, 'Window', {
-							permission,
-							type: contentsType,
-						});
-					}
-					callback(false);
-				}
-			}
-		);
-
-		// Spell check suggestions: Electron renders red squiggles automatically when
-		// `spellcheck` is true on a form element, but the right-click "Did you mean..."
-		// menu has to be wired up in the main process.
-		browserWindow.webContents.on('context-menu', (_event, params) => {
-			logger.debug('context-menu fired', 'Window', {
-				isEditable: params.isEditable,
-				misspelledWord: params.misspelledWord,
-				suggestions: params.dictionarySuggestions,
-				selectionText: params.selectionText,
-			});
-
-			if (!params.isEditable) return;
-
-			const template: Electron.MenuItemConstructorOptions[] = [];
-
-			const suggestions = params.dictionarySuggestions ?? [];
-			if (params.misspelledWord) {
-				if (suggestions.length === 0) {
-					template.push({ label: 'No suggestions', enabled: false });
-				} else {
-					for (const suggestion of suggestions) {
-						template.push({
-							label: suggestion,
-							click: () => browserWindow.webContents.replaceMisspelling(suggestion),
-						});
-					}
-				}
-				template.push(
-					{ type: 'separator' },
-					{
-						label: 'Add to Dictionary',
-						click: () =>
-							browserWindow.webContents.session.addWordToSpellCheckerDictionary(
-								params.misspelledWord
-							),
-					},
-					{ type: 'separator' }
-				);
-			}
-
-			template.push(
-				{ role: 'cut' },
-				{ role: 'copy' },
-				{ role: 'paste' },
-				{ type: 'separator' },
-				{ role: 'selectAll' }
-			);
-
-			Menu.buildFromTemplate(template).popup({ window: browserWindow });
-		});
+		attachSpellCheckContextMenu(browserWindow);
 
 		browserWindow.on('closed', () => {
 			logger.info('Browser window closed', 'Window');
@@ -789,120 +314,7 @@ export function createWindowManager(deps: WindowManagerDependencies): WindowMana
 		// These handlers capture crashes that Sentry in the renderer cannot
 		// report (because the renderer process is dead or broken).
 
-		// Handle renderer process termination (crash, kill, OOM, etc.)
-		browserWindow.webContents.on('render-process-gone', (_event, details) => {
-			logger.error('Renderer process gone', 'Window', {
-				reason: details.reason,
-				exitCode: details.exitCode,
-			});
-
-			// `killed` (signal-terminated, e.g. app quit / OS shutdown / user
-			// force-quit) and `clean-exit` are intentional terminations, not
-			// crashes - the auto-reload guard below already treats them as such.
-			// Reporting them as `fatal` Sentry events is pure noise; genuine
-			// out-of-memory kills surface separately as reason `oom`. Only the
-			// real crash reasons (`crashed`, `oom`, `abnormal-exit`, etc.) are
-			// worth a breadcrumb. Fixes MAESTRO-4X/4Y.
-			const intentionalTermination = details.reason === 'killed' || details.reason === 'clean-exit';
-			if (!intentionalTermination) {
-				// Report to Sentry from main process (always available)
-				reportCrashToSentry(`Renderer process gone: ${details.reason}`, 'fatal', {
-					reason: details.reason,
-					exitCode: details.exitCode,
-				});
-			}
-
-			// Auto-reload unless the process was intentionally killed
-			if (details.reason !== 'killed' && details.reason !== 'clean-exit') {
-				logger.info('Attempting to reload renderer after crash', 'Window');
-				setTimeout(() => {
-					if (!browserWindow.isDestroyed()) {
-						browserWindow.webContents.reload();
-					}
-				}, 1000);
-			}
-		});
-
-		// Handle window becoming unresponsive (frozen renderer)
-		browserWindow.on('unresponsive', () => {
-			logger.warn('Window became unresponsive', 'Window');
-			reportCrashToSentry('Window unresponsive', 'warning', {
-				memoryUsage: process.memoryUsage(),
-			});
-		});
-
-		// Log when window recovers from unresponsive state
-		browserWindow.on('responsive', () => {
-			logger.info('Window became responsive again', 'Window');
-		});
-
-		// Note: the legacy 'crashed' event was removed in Electron 41 and
-		// is now subsumed by 'render-process-gone' above (which reports to
-		// Sentry with full reason/exitCode detail and handles auto-reload).
-
-		// Handle page load failures (network issues, invalid URLs, etc.)
-		browserWindow.webContents.on(
-			'did-fail-load',
-			(_event, errorCode, errorDescription, validatedURL) => {
-				// Ignore aborted loads (user navigated away)
-				if (errorCode === -3) return;
-
-				logger.error('Page failed to load', 'Window', {
-					errorCode,
-					errorDescription,
-					url: validatedURL,
-				});
-				reportCrashToSentry(`Page failed to load: ${errorDescription}`, 'error', {
-					errorCode,
-					errorDescription,
-					url: validatedURL,
-				});
-			}
-		);
-
-		// Handle preload script errors
-		browserWindow.webContents.on('preload-error', (_event, preloadScriptPath, error) => {
-			logger.error('Preload script error', 'Window', {
-				preloadPath: preloadScriptPath,
-				error: error.message,
-				stack: error.stack,
-			});
-			reportCrashToSentry('Preload script error', 'fatal', {
-				preloadPath: preloadScriptPath,
-				error: error.message,
-				stack: error.stack,
-			});
-		});
-
-		// Forward renderer console errors to main process logger and Sentry
-		// This catches errors that happen before or outside React's error boundary
-		browserWindow.webContents.on('console-message', (_event, level, message, line, sourceId) => {
-			// Level 2 = error (0=verbose, 1=info, 2=warning, 3=error)
-			if (level === 3) {
-				logger.error(`Renderer console error: ${message}`, 'Window', {
-					line,
-					source: sourceId,
-				});
-
-				// Report critical errors to Sentry
-				// Filter out common noise (React dev warnings, etc.)
-				const isCritical =
-					message.includes('Uncaught') ||
-					message.includes('TypeError') ||
-					message.includes('ReferenceError') ||
-					message.includes('Cannot read') ||
-					message.includes('is not defined') ||
-					message.includes('is not a function');
-
-				if (isCritical) {
-					reportCrashToSentry(`Renderer error: ${message}`, 'error', {
-						line,
-						source: sourceId,
-					});
-				}
-			}
-		});
-
+		attachWindowCrashHandlers(browserWindow);
 		// Initialize auto-updater (only in production, and only for the primary
 		// window - update checks/installs are a single app-wide concern).
 		if (isMain) {
@@ -1034,41 +446,4 @@ function notifySessionsReclaimedToPrimary(registry: WindowRegistry, count: numbe
 		message: `${count} ${noun} moved to main window`,
 		color: 'theme' as const,
 	});
-}
-
-// Track if stub handlers have been registered (module-level to persist across createWindow calls)
-let devStubsRegistered = false;
-
-/**
- * Registers stub IPC handlers for auto-updater in development mode.
- * These provide helpful error messages instead of silent failures.
- * Uses a module-level flag to ensure handlers are only registered once.
- */
-function registerDevAutoUpdaterStubs(): void {
-	// Only register once - prevents duplicate handler errors if createWindow is called multiple times
-	if (devStubsRegistered) {
-		logger.debug('Auto-updater stub handlers already registered, skipping', 'Window');
-		return;
-	}
-
-	ipcMain.handle('updates:download', async () => {
-		return {
-			success: false,
-			error: 'Auto-update is disabled in development mode. Please check update first.',
-		};
-	});
-
-	ipcMain.handle('updates:install', async () => {
-		logger.warn('Auto-update install called in development mode', 'AutoUpdater');
-	});
-
-	ipcMain.handle('updates:getStatus', async () => {
-		return { status: 'idle' as const };
-	});
-
-	ipcMain.handle('updates:checkAutoUpdater', async () => {
-		return { success: false, error: 'Auto-update is disabled in development mode' };
-	});
-
-	devStubsRegistered = true;
 }
